GUAC-1364: Move session management into AuthenticationService. The token map should not appear anywhere else (except where needed to configure dependency injection of the same).
